backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
distro==1.9.0
Flask==3.0.0
Flask-Cors==4.0.0
groq==0.11.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.6
supabase==2.10.0
PyJWT==2.8.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.114.0
cryptography==42.0.0
